requirements.txt
Flask==2.2.5
Flask-WTF==1.1.1
Werkzeug==2.2.3
Pillow==10.0.0
PyMuPDF==1.23.7
pytesseract==0.3.10
reportlab==4.0.5
